Decompose consumption floor into equiv-param + unequiv-DAG; fix prose
Major refactor splitting the consumption floor into a per-equivalent
estimation parameter (`consumption_equiv_floor`, ~$1.6k/yr) and a
household-$ DAG-derived value
(`consumption_unequiv_floor = consumption_equiv_floor *
equivalence_scale`). Same for consumption: `consumption_unequiv` is
the action variable, `consumption_equiv = consumption_unequiv /
equivalence_scale` is a new DAG function fed into `utility`.

`transfers` and `borrowing_constraint` now consume
`consumption_unequiv_floor` (DAG output) directly — no inline
multiplication. `utility` consumes `consumption_equiv` directly — no
inline division.

`MAX_CONSUMPTION_UNEQUIV` is no longer attached as a dynamic Model
attribute; `inject_consumption_unequiv_points` imports it as a module
constant. The runtime grid's first two points are pinned exactly to
the singles' floor and the married floor (`* 2 ** exponent`), with
geomspace from there to `MAX_CONSUMPTION_UNEQUIV` — so both
household-floor levels land on a feasible action regardless of
sub-ULP drift in `jnp.geomspace`.

Pref-type-indexed Series renamed to plural form
(`consumption_weights`, `coefficients_rra`); the per-cell scalars
(after `[pref_type]` indexing) take the singular form. Drops the
`alpha` / `gamma` aliases. `scaled_bequest_weight` continues to
consume the per-cell scalars (pylcm broadcasts pref-type-indexed
Series to per-cell scalars before consumption).

Also tighten `_build_per_target_*` annotations to
`dict[RegimeName, ...]`, drop stale `consumption_unequiv` prose in
docstrings/comments where bare "consumption" was meant, and update
all callers + tests for the new signatures. The benchmark snapshot
is regenerated against the new keys/structure.

+++ src/aca_model/environment/social_security.py
@@ -30,7 +30,7 @@ def next_claimed_ss(
 
 def enter_claimed_ss() -> DiscreteState:
     """Initial claimed_ss when entering the SS eligibility window."""
-    return ClaimedSS.no
+    return jnp.int32(ClaimedSS.no)
 
 
 # --- PIA lookup (DAG functions) ---
